+++ src/Set/Defer.php
@@ -26,7 +26,7 @@ final class Defer implements Implementation
      */
     public function __construct(Sequence\Implementation $values)
     {
-        $this->values = $values->distinct();
+        $this->values = $values;
     }
 
     /**
@@ -36,7 +36,7 @@ final class Defer implements Implementation
      */
     public function __invoke($element): self
     {
-        return new self(($this->values)($element));
+        return self::distinct(($this->values)($element));
     }
 
     /**
@@ -49,7 +49,7 @@ final class Defer implements Implementation
      */
     public static function of(\Generator $generator): self
     {
-        return new self(new Sequence\Defer($generator));
+        return self::distinct(new Sequence\Defer($generator));
     }
 
     public function size(): int
@@ -101,10 +101,6 @@ final class Defer implements Implementation
      */
     public function remove($element): self
     {
-        if (!$this->contains($element)) {
-            return $this;
-        }
-
         return new self($this->values->filter(
             static fn($value) => $value !== $element,
         ));
@@ -174,7 +170,7 @@ final class Defer implements Implementation
      */
     public function map(callable $function): self
     {
-        return new self($this->values->map($function));
+        return self::distinct($this->values->map($function));
     }
 
     /**
@@ -210,7 +206,7 @@ final class Defer implements Implementation
      */
     public function merge(Implementation $set): self
     {
-        return new self($this->values->append($set->sequence()));
+        return self::distinct($this->values->append($set->sequence()));
     }
 
     /**
@@ -244,10 +240,30 @@ final class Defer implements Implementation
     }
 
     /**
+     * @template R
+     * @param R $carry
+     * @param callable(R, T): R $assert
+     *
+     * @return self<T>
+     */
+    public function safeguard($carry, callable $assert): self
+    {
+        return new self($this->values->safeguard($carry, $assert));
+    }
+
+    /**
      * @return Sequence\Implementation<T>
      */
     public function sequence(): Sequence\Implementation
     {
         return $this->values;
+    }
+
+    /**
+     * @psalm-pure
+     */
+    private static function distinct(Sequence\Implementation $values): self
+    {
+        return new self($values->distinct());
     }
 }