+++ src/Set/Lazy.php
@@ -27,7 +27,7 @@ final class Lazy implements Implementation
      */
     public function __construct(Sequence\Implementation $values)
     {
-        $this->values = $values->distinct();
+        $this->values = $values;
     }
 
     /**
@@ -37,7 +37,7 @@ final class Lazy implements Implementation
      */
     public function __invoke($element): self
     {
-        return new self(($this->values)($element));
+        return self::distinct(($this->values)($element));
     }
 
     /**
@@ -50,7 +50,7 @@ final class Lazy implements Implementation
      */
     public static function of(callable $generator): self
     {
-        return new self(new Sequence\Lazy($generator));
+        return self::distinct(new Sequence\Lazy($generator));
     }
 
     public function size(): int
@@ -102,10 +102,6 @@ final class Lazy implements Implementation
      */
     public function remove($element): self
     {
-        if (!$this->contains($element)) {
-            return $this;
-        }
-
         return new self($this->values->filter(
             static fn($value) => $value !== $element,
         ));
@@ -175,7 +171,7 @@ final class Lazy implements Implementation
      */
     public function map(callable $function): self
     {
-        return new self($this->values->map($function));
+        return self::distinct($this->values->map($function));
     }
 
     /**
@@ -211,7 +207,7 @@ final class Lazy implements Implementation
      */
     public function merge(Implementation $set): self
     {
-        return new self($this->values->append($set->sequence()));
+        return self::distinct($this->values->append($set->sequence()));
     }
 
     /**
@@ -245,10 +241,30 @@ final class Lazy implements Implementation
     }
 
     /**
+     * @template R
+     * @param R $carry
+     * @param callable(R, T): R $assert
+     *
+     * @return self<T>
+     */
+    public function safeguard($carry, callable $assert): self
+    {
+        return new self($this->values->safeguard($carry, $assert));
+    }
+
+    /**
      * @return Sequence\Implementation<T>
      */
     public function sequence(): Sequence\Implementation
     {
         return $this->values;
+    }
+
+    /**
+     * @psalm-pure
+     */
+    private static function distinct(Sequence\Implementation $values): self
+    {
+        return new self($values->distinct());
     }
 }